vistoria/requirements.txt
# Sistema Ágil - Vistoria de Veículos
# Dependências do projeto

# Flask e suas extensões
Flask==3.0.0
Flask-CORS==4.0.0

# Banco de dados PostgreSQL
psycopg2==2.9.9
psycopg2-binary==2.9.9

# Processamento de imagens
Pillow==10.1.0

# Utilitários
python-dateutil==2.8.2
python-dotenv==1.0.0

# Para desenvolvimento/debug (opcional)
werkzeug==3.0.1

# Validação e serialização (opcional)
marshmallow==3.20.1

# Logs estruturados (opcional)
structlog==23.2.0

# Configuração de ambiente
python-decouple==3.8
